Decimal value fix for POS

+++ src/main/java/com/unisys/trans/cps/middleware/models/response/POSResponseDTO.java
@@ -7,7 +7,7 @@ public class POSResponseDTO {
     
     private String valueType;
 
-    private Long value;
+    private float value;
 
     private String unit;
